+++ bpm-engine-impl/src/main/java/io/takari/bpm/EngineListener.java
@@ -1,0 +1,8 @@
+package io.takari.bpm;
+
+import io.takari.bpm.state.ProcessInstance;
+
+public interface EngineListener {
+
+    ProcessInstance onFinalize(ProcessInstance state);
+}